move debug impls into their own module
